chore: move `boards` realm to `boards-original`
This is the first step on keeping the history explicit in GitHub when
blaming. Using a second branch `boards` will be moved to `boards2` also
keeping the history, and finally `boards-original` will be moved back to
`boards`. This is a hacky way to keep GitHub history explicit for all
authors.